requirements.txt
python-telegram-bot==21.6
anthropic==0.40.0
groq>=1.1.2
httpx>=0.27.0
